Non-ui delegation.rs
