+++ src/main.rs
@@ -11,6 +11,7 @@ mod about;
 mod comms;
 mod date_ago;
 mod db;
+mod delegation;
 mod error;
 mod events;
 mod feed;
